# src/main/java/br/com/fiap3esa/autoescola3esa/domain/instrutor/DadosListagemInstrutor.java
package br.com.fiap3esa.autoescola3esa.domain.instrutor;

public record DadosListagemInstrutor(
        Long id,
        String nome,
        String email,
        Especialidade especialidade) {
    public DadosListagemInstrutor(Instrutor instrutor) {
        this(
                instrutor.getId(),
                instrutor.getNome(),
                instrutor.getEmail(),
                instrutor.getEspecialidade()
        );
    }
}