# src/main/java/br/com/fiap3esa/autoescola3esa/domain/usuario/DadosLogin.java
package br.com.fiap3esa.autoescola3esa.domain.usuario;

import jakarta.validation.constraints.NotBlank;

public record DadosLogin(
        @NotBlank
        String login,

        @NotBlank
        String senha) {
}